+++ app/src/main/java/es/developers/achambi/afines/invoices/presenter/InvoiceDetailsPresenter.kt
@@ -22,7 +22,7 @@ class InvoiceDetailsPresenter(invoiceDetailsScreen: InvoiceDetailsScreen,
                               private val presentationBuilder: InvoiceDetailsPresentationBuilder)
     : Presenter<InvoiceDetailsScreen>(invoiceDetailsScreen, lifecycle, executor) {
 
-    fun onViewCreated(invoiceId: Int) {
+    fun onViewCreated(invoiceId: Long) {
         screen.showDetailsLoading()
         val responseHandler = object : ResponseHandler<Invoice?> {
             override fun onSuccess(response: Invoice?) {
@@ -46,7 +46,7 @@ class InvoiceDetailsPresenter(invoiceDetailsScreen: InvoiceDetailsScreen,
         request(request, responseHandler)
     }
 
-    fun onUserDownloadClicked(invoiceId: Int) {
+    fun onUserDownloadClicked(invoiceId: Long) {
         screen.showDownloadInProgress()
         val responseHandler = object : ResponseHandler<DetailedInvoice?> {
             override fun onSuccess(response: DetailedInvoice?) {
@@ -63,7 +63,7 @@ class InvoiceDetailsPresenter(invoiceDetailsScreen: InvoiceDetailsScreen,
         request(request, responseHandler)
     }
 
-    fun onUserDeleteSelected(invoiceId: Int) {
+    fun onUserDeleteSelected(invoiceId: Long) {
         val responseHandler = object : ResponseHandler<Any> {
             override fun onSuccess(response: Any) {
                 screen.showInvoiceDeleted()
@@ -80,7 +80,7 @@ class InvoiceDetailsPresenter(invoiceDetailsScreen: InvoiceDetailsScreen,
         request(request, responseHandler)
     }
 
-    fun onUserFileBytesRequired(invoiceId: Int, uri: Uri?, activity: Activity?) {
+    fun onUserFileBytesRequired(invoiceId: Long, uri: Uri?, activity: Activity?) {
         val responseHandler = object : ResponseHandler<ByteArray?> {
             override fun onSuccess(response: ByteArray?) {
                 if(response != null && uri != null) {
